Common tests now work...
