+++ test/extensions/filters/network/common/redis/mocks.h
@@ -5,6 +5,7 @@
 #include <string>
 
 #include "extensions/filters/network/common/redis/codec_impl.h"
+#include "extensions/filters/network/common/redis/client_impl.h"
 
 #include "test/test_common/printers.h"
 
@@ -23,6 +24,80 @@ namespace Redis {
 void PrintTo(const RespValue& value, std::ostream* os);
 void PrintTo(const RespValuePtr& value, std::ostream* os);
 bool operator==(const RespValue& lhs, const RespValue& rhs);
+
+class MockEncoder : public Common::Redis::Encoder {
+public:
+  MockEncoder();
+  ~MockEncoder();
+
+  MOCK_METHOD2(encode, void(const Common::Redis::RespValue& value, Buffer::Instance& out));
+
+private:
+  Common::Redis::EncoderImpl real_encoder_;
+};
+
+class MockDecoder : public Common::Redis::Decoder {
+public:
+  MockDecoder();
+  ~MockDecoder();
+
+  MOCK_METHOD1(decode, void(Buffer::Instance& data));
+};
+
+namespace Client {
+
+class MockClient : public Client {
+public:
+  MockClient();
+  ~MockClient();
+
+  void raiseEvent(Network::ConnectionEvent event) {
+    for (Network::ConnectionCallbacks* callbacks : callbacks_) {
+      callbacks->onEvent(event);
+    }
+  }
+
+  void runHighWatermarkCallbacks() {
+    for (auto* callback : callbacks_) {
+      callback->onAboveWriteBufferHighWatermark();
+    }
+  }
+
+  void runLowWatermarkCallbacks() {
+    for (auto* callback : callbacks_) {
+      callback->onBelowWriteBufferLowWatermark();
+    }
+  }
+
+  MOCK_METHOD1(addConnectionCallbacks, void(Network::ConnectionCallbacks& callbacks));
+  MOCK_METHOD0(close, void());
+  MOCK_METHOD2(makeRequest, PoolRequest*(const Common::Redis::RespValue& request,
+                                                        PoolCallbacks& callbacks));
+
+  std::list<Network::ConnectionCallbacks*> callbacks_;
+};
+
+class MockPoolRequest : public PoolRequest {
+public:
+  MockPoolRequest();
+  ~MockPoolRequest();
+
+  MOCK_METHOD0(cancel, void());
+};
+
+class MockPoolCallbacks : public PoolCallbacks {
+public:
+  MockPoolCallbacks();
+  ~MockPoolCallbacks();
+
+  void onResponse(Common::Redis::RespValuePtr&& value) override { onResponse_(value); }
+
+  MOCK_METHOD1(onResponse_, void(Common::Redis::RespValuePtr& value));
+  MOCK_METHOD0(onFailure, void());
+};
+
+
+} // namespace Client
 
 } // namespace Redis
 } // namespace Common
