+++ test/extensions/filters/network/redis_proxy/mocks.h
@@ -18,85 +18,16 @@ namespace Extensions {
 namespace NetworkFilters {
 namespace RedisProxy {
 
-class MockEncoder : public Common::Redis::Encoder {
-public:
-  MockEncoder();
-  ~MockEncoder();
-
-  MOCK_METHOD2(encode, void(const Common::Redis::RespValue& value, Buffer::Instance& out));
-
-private:
-  Common::Redis::EncoderImpl real_encoder_;
-};
-
-class MockDecoder : public Common::Redis::Decoder {
-public:
-  MockDecoder();
-  ~MockDecoder();
-
-  MOCK_METHOD1(decode, void(Buffer::Instance& data));
-};
-
 namespace ConnPool {
-
-class MockClient : public Common::Redis::Client {
-public:
-  MockClient();
-  ~MockClient();
-
-  void raiseEvent(Network::ConnectionEvent event) {
-    for (Network::ConnectionCallbacks* callbacks : callbacks_) {
-      callbacks->onEvent(event);
-    }
-  }
-
-  void runHighWatermarkCallbacks() {
-    for (auto* callback : callbacks_) {
-      callback->onAboveWriteBufferHighWatermark();
-    }
-  }
-
-  void runLowWatermarkCallbacks() {
-    for (auto* callback : callbacks_) {
-      callback->onBelowWriteBufferLowWatermark();
-    }
-  }
-
-  MOCK_METHOD1(addConnectionCallbacks, void(Network::ConnectionCallbacks& callbacks));
-  MOCK_METHOD0(close, void());
-  MOCK_METHOD2(makeRequest, Common::Redis::PoolRequest*(const Common::Redis::RespValue& request,
-                                                        Common::Redis::PoolCallbacks& callbacks));
-
-  std::list<Network::ConnectionCallbacks*> callbacks_;
-};
-
-class MockPoolRequest : public Common::Redis::PoolRequest {
-public:
-  MockPoolRequest();
-  ~MockPoolRequest();
-
-  MOCK_METHOD0(cancel, void());
-};
-
-class MockPoolCallbacks : public Common::Redis::PoolCallbacks {
-public:
-  MockPoolCallbacks();
-  ~MockPoolCallbacks();
-
-  void onResponse(Common::Redis::RespValuePtr&& value) override { onResponse_(value); }
-
-  MOCK_METHOD1(onResponse_, void(Common::Redis::RespValuePtr& value));
-  MOCK_METHOD0(onFailure, void());
-};
 
 class MockInstance : public Instance {
 public:
   MockInstance();
   ~MockInstance();
 
-  MOCK_METHOD3(makeRequest, Common::Redis::PoolRequest*(const std::string& hash_key,
+  MOCK_METHOD3(makeRequest, Common::Redis::Client::PoolRequest*(const std::string& hash_key,
                                                         const Common::Redis::RespValue& request,
-                                                        Common::Redis::PoolCallbacks& callbacks));
+                                                        Common::Redis::Client::PoolCallbacks& callbacks));
 };
 
 } // namespace ConnPool
